+++ index.tsx
@@ -1,4 +1,6 @@
 export { Breadcrumbs } from "./src/Breadcrumbs";
 export { FolderStructure } from "./src/FolderStructure";
+export { PageScrollIndicator } from "./src/PageScrollIndicator";
+export { Rating } from "./src/Rating";
 export { SkipLink } from "./src/SkipLink";
 export { Slider } from "./src/Slider";